hex/requirements.txt
pytest
pygame-ce
pyyaml
webcolors
numpy